feat(components): разбиение кода на модули
